improvement(tester.py): move out ldap code to ldap.py
